+++ examples/common/app.h
@@ -4,4 +4,4 @@
 
 bool app_init    ();
 void app_shutdown();
-void app_render  (double t, hmm_mat4 view, hmm_mat4 proj);
+void app_render  (float t, hmm_mat4 view, hmm_mat4 proj);
